Control de acceso implementado para Post tanto en web como en API. Tambien se controla de cara al HTML

+++ src/main/java/es/codeurjc/web/controller/PostController.java
@@ -26,7 +26,6 @@ import es.codeurjc.web.dto.PostDTO;
 import es.codeurjc.web.service.CommentService;
 import es.codeurjc.web.service.PostService;
 import es.codeurjc.web.service.SectionService;
-import es.codeurjc.web.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 
 @Controller
@@ -40,8 +39,6 @@ public class PostController {
 
     @Autowired
     private SectionService sectionService;
-    @Autowired
-    private UserService userService;
 
     @GetMapping("/post")
     public String viewPosts(Model model, @RequestParam(defaultValue = "0") int page) {
@@ -72,13 +69,16 @@ public class PostController {
 
     @PostMapping("/post/new")
     public String createPost(Model model, CreatePostDTO createPostDTO, @RequestParam MultipartFile newImage, @RequestParam String newContributors, @RequestParam(value = "sections", required = false) List<Long> sectionIds, HttpServletRequest request) throws IOException {
-
+        if (createPostDTO.title().isEmpty()) {
+            model.addAttribute("message", "The title cannot be empty");
+            return "error";
+        }
         postService.addSections(createPostDTO, sectionIds);
 
         String[] contributorsArray = newContributors.split(",");
         postService.addContributors(createPostDTO, contributorsArray);
         postService.save(createPostDTO, newImage, request);
-        return "redirect:https://localhost:8443/post";
+        return "redirect:/post";
     }
 
     @GetMapping("/post/{postId}")
@@ -87,23 +87,26 @@ public class PostController {
 
 
         if (request.getUserPrincipal() != null) {
+
             model.addAttribute("post", postDTO);
             model.addAttribute("comments", commentService.findAllCommentsByPostId(postId));
-           
             model.addAttribute("currentPage", 0); //commentPage.getNumber());
             model.addAttribute("hasImage", postDTO.image() != null);
+            model.addAttribute("isOwner", postService.checkIfUserIsTheOwner(postId, request));
             model.addAttribute("logged", true);
 
             return "view_post";
 
         } else {
+
             model.addAttribute("post", postDTO);
             model.addAttribute("comments", commentService.findAllCommentsByPostId(postId));
             model.addAttribute("currentPage", 0); //commentPage.getNumber());
             model.addAttribute("hasImage", postDTO.image() != null);
-
             model.addAttribute("logged", false);
+            
             return "view_post";
+        
         }
 
     }
@@ -116,50 +119,61 @@ public class PostController {
     }
 
     @GetMapping("/post/{postId}/edit")
-    public String updatePost(Model model, @PathVariable Long postId) {
+    public String updatePost(Model model, @PathVariable Long postId, HttpServletRequest request) {
 
-        PostDTO postDTO = postService.findByIdAsDTO(postId);
+        if (postService.checkIfUserIsTheOwner(postId, request)) {
+            
+            PostDTO postDTO = postService.findByIdAsDTO(postId);
 
-        List<Map<String, Object>> markedSections = postService.preparePostSectionsForForm(postId);
+            List<Map<String, Object>> markedSections = postService.preparePostSectionsForForm(postId);
 
-        // Create a List<Section> with "selected" property
-        // List<Map<String, Object>> sectionsWithSelection = new ArrayList<>();
-        // for (Section section : allSections) {
-        //     Map<String, Object> sectionData = new HashMap<>();
-        //     sectionData.put("postId", section.getId());
-        //     sectionData.put("title", section.getTitle());
-        //     // Verify if the section is in the post using its postId
-        //     boolean isSelected = postSections.stream()
-        //             .anyMatch(s -> s.getId() == section.getId());
-        //     sectionData.put("selected", isSelected);
-        //     sectionsWithSelection.add(sectionData);
-        // }
-        String contributors = postService.contributorsToString(postId);
+            String contributors = postService.contributorsToString(postId);
 
-        model.addAttribute("sections", markedSections);
-        model.addAttribute("post", postDTO);
-        model.addAttribute("title", postDTO.title());
-        model.addAttribute("content", postDTO.content());
-        model.addAttribute("contributors", contributors);
-        model.addAttribute("isEditing", true);
-        return "post_form";
+            model.addAttribute("sections", markedSections);
+            model.addAttribute("post", postDTO);
+            model.addAttribute("title", postDTO.title());
+            model.addAttribute("content", postDTO.content());
+            model.addAttribute("contributors", contributors);
+            model.addAttribute("isEditing", true);
+            
+            return "post_form";
+
+        } else {
+            model.addAttribute("message", "You are not allowed to do this");
+            return "error";
+        }
 
     }
 
     @PostMapping("/post/{postId}/edit")
     public String updatePost(Model model, @PathVariable Long postId, CreatePostDTO createPostDTO,
-            @RequestAttribute MultipartFile newImage, @RequestParam(value = "sections", required = false) List<Long> newSectionIds, @RequestParam("newContributors") String newContributorsStrings)
-            throws IOException {
+            @RequestAttribute MultipartFile newImage, @RequestParam(value = "sections", required = false) List<Long> newSectionIds,
+            @RequestParam("newContributors") String newContributorsStrings, HttpServletRequest request) throws IOException {
 
-        postService.updatePost(postId, createPostDTO, newSectionIds, newContributorsStrings.split(","), newImage);
-        return "redirect:/post/" + postId;
+        if (postService.checkIfUserIsTheOwner(postId, request)) {
+
+            postService.updatePost(postId, createPostDTO, newSectionIds, newContributorsStrings.split(","), newImage);
+            return "redirect:/post/" + postId;
+
+        } else {
+            model.addAttribute("message", "You are not allowed to do this");
+            return "error";
+        }
 
     }
 
     @PostMapping("/post/{postId}/delete")
-    public String deletePost(@PathVariable Long postId, Model model) {
-        postService.deletePost(postId);
-        return "redirect:/post";
+    public String deletePost(@PathVariable Long postId, Model model, HttpServletRequest request) {
+        
+        if (postService.checkIfUserIsTheOwner(postId, request)) {
+        
+            postService.deletePost(postId);
+            return "redirect:/post";
+
+        } else {
+            model.addAttribute("message", "You are not allowed to do this");
+            return "error";
+        }
 
     }
 
@@ -172,6 +186,7 @@ public class PostController {
 
     @PostMapping("/post/{postId}/comment/new")
     public String newPostComment(Model model, @PathVariable Long postId, CreateCommentDTO newComment, HttpServletRequest request) {
+        
         if (postService.existsById(postId)) {
 
             if (newComment.content().isEmpty()) {
@@ -197,10 +212,13 @@ public class PostController {
     public String editPostComment(@PathVariable Long postId, @PathVariable Long commentId, Model model, HttpServletRequest request) {
 
         if (commentService.checkIfCommentOwnerAndCommnetOnPost(request, postId, commentId)) {
+            
             model.addAttribute("post", postService.findByIdAsDTO(postId));
             model.addAttribute("comment", commentService.findCommentByIdDTO(commentId));
             model.addAttribute("isEditing", true);
+            
             return "comment_form";
+
         } else {
             model.addAttribute("message", "No tienes permisos para realizar esa acción, o estás intentando editar un comentario de otro post");
             return "error";
@@ -213,6 +231,7 @@ public class PostController {
             CommentDTO updatedComment, HttpServletRequest request) {
 
         if (commentService.checkIfCommentOwnerAndCommnetOnPost(request, postId, commentId)) {
+            
             postService.findByIdAsDTO(postId);
             commentService.findCommentByIdDTO(commentId);
 
@@ -224,7 +243,9 @@ public class PostController {
                 model.addAttribute("message", "El valor del rating debe estar entre 0 y 5");
                 return "error";
             }
+
             commentService.updateComment(commentId, updatedComment, postId);
+            
             return "redirect:/post/" + postId;
 
         } else {
@@ -235,16 +256,19 @@ public class PostController {
     }
 
     @PostMapping("/post/{postId}/comment/{commentId}/delete")
-    public String deletePostComment(@PathVariable Long postId, @PathVariable Long commentId, Model model, HttpServletRequest request
-    ) {
+    public String deletePostComment(@PathVariable Long postId, @PathVariable Long commentId, Model model, HttpServletRequest request) {
+        
         if (commentService.checkIfCommentOwnerAndCommnetOnPost(request, postId, commentId)) {
+        
             PostDTO postDTO = postService.findByIdAsDTO(postId);
             commentService.findCommentByIdDTO(commentId);
-
             commentService.deleteCommentFromPost(postId, commentId);
+            
             model.addAttribute("post", postDTO);
             model.addAttribute("Comments", postDTO.comments());
+        
             return "redirect:/post/" + postId;
+        
         } else {
             model.addAttribute("message", "No tienes permisos para realizar esa acción, o estás intentando eliminar un comentario de otro post");
             return "error";
